make ThemeChoice inner class of Config, make theme instances fields of ThemeChoice, move isDarkLaf back to ThemeUtil
