--- enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/ThemeChoice.java
@@ -1,63 +1,0 @@
-package org.quiltmc.enigma.gui.config.theme;
-
-import org.quiltmc.config.api.Config;
-import org.quiltmc.config.api.values.ComplexConfigValue;
-import org.quiltmc.config.api.values.ConfigSerializableObject;
-import org.quiltmc.enigma.gui.config.theme.properties.composite.SyntaxPaneProperties;
-
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.image.BufferedImage;
-import javax.swing.JPanel;
-
-public enum ThemeChoice implements ConfigSerializableObject<String> {
-	DEFAULT(
-	),
-	DARCULA(
-	),
-	DARCERULA(
-	),
-	METAL(
-	),
-	SYSTEM(
-	),
-	NONE(
-	);
-
-
-	private static void testColorBuilder(Config.SectionBuilder builder) {
-		final var colors = new SyntaxPaneProperties.Colors.Builder().build();
-		colors.stream().forEach(builder::field);
-	}
-
-	public static boolean isDarkLaf() {
-		// a bit of a hack because swing doesn't give any API for that, and we need colors that aren't defined in look and feel
-		JPanel panel = new JPanel();
-		panel.setSize(new Dimension(10, 10));
-		panel.doLayout();
-
-		BufferedImage image = new BufferedImage(panel.getSize().width, panel.getSize().height, BufferedImage.TYPE_INT_RGB);
-		panel.printAll(image.getGraphics());
-
-		Color c = new Color(image.getRGB(0, 0));
-
-		// convert the color we got to grayscale
-		int b = (int) (0.3 * c.getRed() + 0.59 * c.getGreen() + 0.11 * c.getBlue());
-		return b < 85;
-	}
-
-	@Override
-	public ThemeChoice convertFrom(String representation) {
-		return ThemeChoice.valueOf(representation);
-	}
-
-	@Override
-	public String getRepresentation() {
-		return this.name();
-	}
-
-	@Override
-	public ComplexConfigValue copy() {
-		return this;
-	}
-}
