+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/properties/MetalThemeProperties.java
@@ -1,6 +1,6 @@
 package org.quiltmc.enigma.gui.config.theme.properties;
 
-import org.quiltmc.enigma.gui.config.theme.ThemeChoice;
+import org.quiltmc.enigma.gui.config.Config;
 import org.quiltmc.enigma.gui.config.theme.properties.composite.ConfigurableConfigCreator;
 import org.quiltmc.enigma.gui.config.theme.properties.composite.SyntaxPaneProperties;
 
@@ -16,11 +16,6 @@ public class MetalThemeProperties extends NonSystemLafThemeProperties {
 
 	protected MetalThemeProperties(SyntaxPaneProperties syntaxPaneColors, List<ConfigurableConfigCreator> creators) {
 		super(syntaxPaneColors, creators);
-	}
-
-	@Override
-	public ThemeChoice getThemeChoice() {
-		return ThemeChoice.METAL;
 	}
 
 	@Override
