+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/properties/SystemThemeProperties.java
@@ -1,6 +1,6 @@
 package org.quiltmc.enigma.gui.config.theme.properties;
 
-import org.quiltmc.enigma.gui.config.theme.ThemeChoice;
+import org.quiltmc.enigma.gui.config.Config;
 import org.quiltmc.enigma.gui.config.theme.properties.composite.ConfigurableConfigCreator;
 import org.quiltmc.enigma.gui.config.theme.properties.composite.SyntaxPaneProperties;
 
@@ -15,11 +15,6 @@ public class SystemThemeProperties extends ThemeProperties {
 
 	protected SystemThemeProperties(SyntaxPaneProperties syntaxPaneColors, List<ConfigurableConfigCreator> creators) {
 		super(syntaxPaneColors, creators);
-	}
-
-	@Override
-	public ThemeChoice getThemeChoice() {
-		return ThemeChoice.SYSTEM;
 	}
 
 	@Override
