+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/util/GuiUtil.java
@@ -3,11 +3,11 @@ package org.quiltmc.enigma.gui.util;
 import com.formdev.flatlaf.extras.FlatSVGIcon;
 import org.quiltmc.enigma.api.analysis.index.jar.EntryIndex;
 import org.quiltmc.enigma.gui.Gui;
-import org.quiltmc.enigma.gui.config.theme.ThemeChoice;
 import org.quiltmc.enigma.api.stats.ProjectStatsResult;
 import org.quiltmc.enigma.api.translation.representation.AccessFlags;
 import org.quiltmc.enigma.api.translation.representation.entry.ClassEntry;
 import org.quiltmc.enigma.api.translation.representation.entry.MethodEntry;
+import org.quiltmc.enigma.gui.config.theme.ThemeUtil;
 import org.quiltmc.enigma.util.Os;
 
 import javax.swing.Icon;
@@ -268,10 +268,10 @@ public class GuiUtil {
 	}
 
 	public static Icon getUpChevron() {
-		return ThemeChoice.isDarkLaf() ? CHEVRON_UP_WHITE : CHEVRON_UP_BLACK;
+		return ThemeUtil.isDarkLaf() ? CHEVRON_UP_WHITE : CHEVRON_UP_BLACK;
 	}
 
 	public static Icon getDownChevron() {
-		return ThemeChoice.isDarkLaf() ? CHEVRON_DOWN_WHITE : CHEVRON_DOWN_BLACK;
+		return ThemeUtil.isDarkLaf() ? CHEVRON_DOWN_WHITE : CHEVRON_DOWN_BLACK;
 	}
 }
