--- enigma-swing/src/test/java/org/quiltmc/enigma/gui/config/ConfigTest.java
@@ -1,17 +1,0 @@
-package org.quiltmc.enigma.gui.config;
-
-import org.junit.jupiter.api.Test;
-import org.quiltmc.enigma.gui.config.theme.ThemeChoice;
-
-import java.util.Arrays;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-public class ConfigTest {
-	@Test
-	void testAllThemeChoicesMapped() {
-		assertTrue(
-			Arrays.stream(ThemeChoice.values()).allMatch(choice -> Config.THEMES_BY_CHOICE.get(choice) != null)
-		);
-	}
-}
